path mismatches between fed and arch

+++ install/config/theme.sh
@@ -4,10 +4,14 @@ gsettings set org.gnome.desktop.interface gtk-theme "Adwaita-dark"
 gsettings set org.gnome.desktop.interface color-scheme "prefer-dark"
 gsettings set org.gnome.desktop.interface icon-theme "Yaru-blue"
 
-# Set links for Nautilius action icons
-sudo ln -snf /usr/share/icons/Adwaita/symbolic/actions/go-previous-symbolic.svg /usr/share/icons/Yaru/scalable/actions/go-previous-symbolic.svg
-sudo ln -snf /usr/share/icons/Adwaita/symbolic/actions/go-next-symbolic.svg /usr/share/icons/Yaru/scalable/actions/go-next-symbolic.svg
-sudo gtk-update-icon-cache /usr/share/icons/Yaru
+# Set links for Nautilus action icons (FEDARCHY: Check if paths exist first)
+if [ -f "/usr/share/icons/Adwaita/symbolic/actions/go-previous-symbolic.svg" ] && [ -d "/usr/share/icons/Yaru/scalable/actions" ]; then
+    sudo ln -snf /usr/share/icons/Adwaita/symbolic/actions/go-previous-symbolic.svg /usr/share/icons/Yaru/scalable/actions/go-previous-symbolic.svg
+    sudo ln -snf /usr/share/icons/Adwaita/symbolic/actions/go-next-symbolic.svg /usr/share/icons/Yaru/scalable/actions/go-next-symbolic.svg
+    sudo gtk-update-icon-cache /usr/share/icons/Yaru
+else
+    echo "Skipping Yaru icon fixes - paths not found on this system"
+fi
 
 # Setup theme links
 mkdir -p ~/.config/omarchy/themes
@@ -26,4 +30,3 @@ ln -snf ~/.config/omarchy/current/theme/btop.theme ~/.config/btop/themes/current
 
 mkdir -p ~/.config/mako
 ln -snf ~/.config/omarchy/current/theme/mako.ini ~/.config/mako/config
-
